reorg of Elasticsearch types, into new package
